pyproject.toml
[tool.poetry]
name = "text2aks"
version = "0.1.1a1"
description = "A simple python library to put text into an image."
authors = ["ImanHpr <imanhpr1999@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/imanhpr/text2aks"
homepage = "https://github.com/imanhpr/text2aks"
keywords = ['Pillow' , 'Picture' , 'Text']
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Other Audience",
    "Natural Language :: Persian",
    "Natural Language :: Arabic",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
mypy = "^0.950"
types-Pillow = "^9.0.15"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
